Fix PID whine noise from wheels during vel control
Given a command, the wheels attempt to achieve it using PID
controllers in the firmware. Ideally, the controller closes
the error between the commanded position and the current one.
But actually, this isn't always possible. For example, if the
goal requires the robot to be within an obstacle, the controller
will push against the obstacle, making no progress, and emitting
a whining noise from the wheel motors. Another example, on carpet,
the controller will find it difficult to achieve the goal exactly,
due to the soft flooring. Unable to close the small error, but
trying all the same, the wheel motors emit a noise. Tuning the
PID parameters is no good because what works well on carpet will
overshoot on hardwood, and vice versa.

In this commit, we relax the controller 0.5 sec after the user
stops sending velocity commands, eliminating the wheel noise.
Note that this commit only address wheel noise in velocity mode
and a future change will be needed to address it in position mode.