Stretch ROS2 driver now uses two callback groups
The first, main group, is a ReentrantCallbackGroup, and covers
nearly all callbacks. This means that the multithreaded executor
can execute subcription, service, and action callbacks as it likes.
The second, mutex group, is a MutuallyExclusiveCallbackGroup, and
covers just the main timer in stretch_driver. This main loop handles
all publishers and motion execution. The main loop is in a separate
group so it doesn't get scheduled in overlapping fashion with
itself (which reentrant groups do), while still executing in
parallel with the callbacks in main group.